+++ src/main/java/frc/robot/commands/intake/IntakeCoral.java
@@ -1,11 +1,9 @@
 package frc.robot.commands.intake;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.intake.CoralIntake;
+import frc.robot.subsystems.superstructure.intake.CoralIntake;
 
-/**
- * use while true on the contoller bindings because end stops motors excute starts them
- */
+/** use while true on the contoller bindings because end stops motors excute starts them */
 public class IntakeCoral extends Command {
 
   private CoralIntake coralIntake;
@@ -18,18 +16,17 @@ public class IntakeCoral extends Command {
 
   @Override
   public void execute() {
-    coralIntake.moveRightMotor(-0.3);
-    coralIntake.moveLeftMotor(-0.3);
+    coralIntake.setMotors(-0.3);
   }
 
   @Override
   public boolean isFinished() {
-    //TODO: use whlie true button in bindings
+    // TODO: use whlie true button in bindings
     return false;
   }
 
   @Override
   public void end(boolean interrupted) {
-    coralIntake.stop();
+    coralIntake.stopMotors();
   }
 }
