+++ src/main/java/frc/robot/commands/intake/ScoreCoral1.java
@@ -1,11 +1,9 @@
 package frc.robot.commands.intake;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.intake.CoralIntake;
+import frc.robot.subsystems.superstructure.intake.CoralIntake;
 
-/**
- * use while true on the contoller bindings because end stops motors excute starts them
- */
+/** use while true on the contoller bindings because end stops motors excute starts them */
 public class ScoreCoral1 extends Command {
 
   private CoralIntake coralIntake;
@@ -19,17 +17,17 @@ public class ScoreCoral1 extends Command {
   @Override
   public void execute() {
     // scores on l1
-    coralIntake.moveRightMotor(0.3);
+    coralIntake.setRightMotor(0.3);
   }
 
   @Override
   public boolean isFinished() {
-    //TODO: use whlie true button in bindings
+    // TODO: use whlie true button in bindings
     return false;
   }
 
   @Override
   public void end(boolean interrupted) {
-    coralIntake.stop();
+    coralIntake.stopMotors();
   }
 }
