+++ src/main/java/frc/robot/commands/intake/ScoreCoral2.java
@@ -1,11 +1,9 @@
 package frc.robot.commands.intake;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.intake.CoralIntake;
+import frc.robot.subsystems.superstructure.intake.CoralIntake;
 
-/**
- * use while true on the contoller bindings because end stops motors excute starts them
- */
+/** use while true on the contoller bindings because end stops motors excute starts them */
 public class ScoreCoral2 extends Command {
 
   private CoralIntake coralIntake;
@@ -18,10 +16,10 @@ public class ScoreCoral2 extends Command {
 
   @Override
   public void execute() {
-    //TODO: use whlie true button in bindings
+    // TODO: use whlie true button in bindings
     // score on not l1
-    coralIntake.moveRightMotor(0.3);
-    coralIntake.moveLeftMotor(0.3);
+    coralIntake.setRightMotor(0.3);
+    coralIntake.setLeftMotor(0.3);
   }
 
   @Override
@@ -31,6 +29,6 @@ public class ScoreCoral2 extends Command {
 
   @Override
   public void end(boolean interrupted) {
-    coralIntake.stop();
+    coralIntake.stopMotors();
   }
 }
